Added: Validation and Exception

+++ src/main/java/com/example/demo/service/AddressBookService.java
@@ -1,45 +1,62 @@
 package com.example.demo.service;
 
-
 import com.example.demo.dto.AddressBookDTO;
+import com.example.demo.exception.AddressBookException;
 import com.example.demo.model.AddressBook;
 import com.example.demo.repository.AddressBookRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-
 import java.util.List;
 import java.util.Optional;
 
 @Service
-public class AddressBookService {
+
+public class AddressBookService implements IAddressBookService {
 
     @Autowired
     AddressBookRepository repository;
 
-
-    public List<AddressBook> getListOfAddresses() {
-        List<AddressBook> addressBooks = repository.findAll();
-        return addressBooks;
+    public String getWelcome() {
+        return "Welcome to Address Book System !!!";
     }
 
-    public AddressBook saveAddress(AddressBookDTO addressBookDTO) {
-        AddressBook addressBook = new AddressBook(addressBookDTO);
-        repository.save(addressBook);
-        return addressBook;
+    public AddressBook postDataToRepo(AddressBookDTO addressBookDTO) {
+        AddressBook  newAddressBook  = new AddressBook (addressBookDTO);
+        repository.save(newAddressBook);
+        return newAddressBook ;
     }
 
-    public AddressBook getAddressbyId(Integer id) {
-        Optional<AddressBook> addressBook = repository.findById(id);
-        return addressBook.get();
+
+    public List<AddressBook > getAllData() {
+        List<AddressBook > list = repository.findAll();
+        return list;
     }
 
-    public AddressBook updateDateById(Integer id, AddressBookDTO addressBookDTO) {
-        AddressBook addressBook = new AddressBook(id, addressBookDTO);
-        repository.save(addressBook);
-        return addressBook;
+    public AddressBook getDataById(Integer id) {
+        Optional<AddressBook> newEmployee = repository.findById(id);
+        if (newEmployee.isPresent()) {
+            return newEmployee.get();
+        } else throw new AddressBookException("Person id not found");
     }
 
-    public void deleteContact(Integer id) {
-        repository.deleteById(id);
+    public AddressBook updateDataById(Integer id, AddressBookDTO addressBookDTO) {
+        Optional<AddressBook> newEmployee = repository.findById(id);
+        if (newEmployee.isPresent()) {
+            AddressBook addressBook = new AddressBook(addressBookDTO);
+            repository.save(addressBook);
+            return addressBook;
+        } else {
+            throw new AddressBookException("Person Not found");
+        }
+    }
+
+    public String deleteDataById(Integer id) {
+        Optional<AddressBook> newEmployee = repository.findById(id);
+        if (newEmployee.isPresent()) {
+            repository.deleteById(id);
+        } else {
+            throw new AddressBookException("Person Details not found");
+        }
+        return null;
     }
 }
